requirements.txt
astroid==1.6.5
blinker==1.4
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Mail==0.9.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.15.4
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyJWT==1.7.1
pylint==1.9.4
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
twilio==6.23.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.0
